move methods to separate package
